rank 2

+++ Level2_my/union/re.c
@@ -3,27 +3,26 @@
 int main (int ac, char **av)
 {
     int i = 0;
-
     int tab [256] = {0};
 
     if (ac == 3)
     {
         while (av[1][i])
         {
-            if (tab[(int)av[1][i]] == 0)
+            if (tab[av[1][i]] == 0)
             {
-                write (1, &av[1][i], 1);
-                tab[(int)av[1][i]] = 1;
+                write(1, &av[1][i], 1);
+                tab[av[1][i]] = 1;
             }
             i++;
         }
         i = 0;
         while (av[2][i])
         {
-            if (tab[(int)av[2][i]] == 0)
+            if (tab[av[2][i]] == 0)
             {
-                write (1, &av[2][i], 1);
-                tab[(int)av[2][i]] = 1;
+                write(1, &av[2][i], 1);
+                tab[av[2][i]] = 1;
             }
             i++;
         }
